+++ src/pages/ProductPage/index.tsx
@@ -1,9 +1,13 @@
-import { ProductCard } from '@/components/ProductCard';
+import { Product } from '@/components/Product';
 import { FC } from 'react';
 import { useParams } from 'react-router-dom';
 
 export const ProductPage: FC = () => {
   const { id } = useParams();
   const productId = Number(id);
-  return <ProductCard id={productId} />;
+  return (
+    <>
+      <Product id={productId} />
+    </>
+  );
 };
