+++ src/theme/styles/components/StyledProductCard.tsx
@@ -1,12 +1,16 @@
+import Slider from 'react-slick';
 import { styled } from 'styled-components';
 import tw from 'twin.macro';
 
+export const StyledProductCardSection = styled.div`
+  ${tw`mt-[160px]`}
+`;
 export const StyledProductCard = styled.div`
   ${tw`grid grid-rows-1 gap-y-6 sm:grid-cols-[41% auto] sm:gap-x-32 mt-8 sm:mt-14 mb-14 overflow-hidden`}
 `;
 
 export const StyledCardRight = styled.div`
-  ${tw`relative flex flex-col h-[290px] lg:h-[580px] bg-main-red pb-8 lg:pb-16 pt-8 rounded-3xl`}
+  ${tw`relative flex flex-col h-[290px] lg:h-[580px] bg-main-white pb-8 lg:pb-16 pt-8 rounded-3xl`}
 `;
 export const StyledSale = styled.div`
   ${tw`w-[191px] flex items-center justify-center py-4 px-16 text-main-white bg-main-red rounded-r`}
@@ -20,13 +24,63 @@ export const StyledNoSale = styled.div`
 `;
 
 export const StyledCardImg = styled.div`
-  ${tw`shrink-0 px-10 lg:px-20 bg-main-black`}
+  ${tw`shrink-0 px-10 lg:px-20 bg-main-white`}
 
   img {
     ${tw`object-contain	h-[211px] lg:w-full lg:h-full`}
   }
 `;
+export const StyledContainerImagesSlider = styled.div`
+${tw`w-full m-auto`}
 
+  .slick-slider {
+    ${tw`relative`}
+    
+    .button {
+      ${tw`absolute top-1/2 -left-14 translate-y-2/4`}
+    }
+
+    .button-next {
+      ${tw`absolute -right-28 left-auto`}
+    }
+
+    .MuiIconButton-root {
+      ${tw`w-6 h-6 p-0`}
+
+      svg {
+        ${tw`stroke-secondary-gray`}
+      }
+
+      &:hover {
+      
+      svg {
+        ${tw`stroke-main-red`}
+      }
+    }
+  }
+
+  .slick-list {
+    ${tw`overflow-hidden`}
+  }
+  .slick-track {
+    ${tw`flex gap-x-5`}
+  }
+  `;
+
+export const StyledImagesSlider = styled(Slider)`
+  .slick-slide {
+    
+      ${tw`w-[500px] `}
+    }
+
+  ${StyledCardImg} {
+    ${tw`min-w-[500px]`}
+  }
+
+  .button {
+    ${tw`hidden lg:inline-flex`}
+  }
+`;
 export const StyledCardInfo = styled.div`
   ${tw`flex flex-col gap-y-2 h-full `}
 `;
